+++ src/main/kotlin/no/njoh/pulseengine/modules/gui/Scrollable.kt
@@ -2,7 +2,7 @@ package no.njoh.pulseengine.modules.gui
 
 sealed interface Scrollable
 {
-    fun onScroll(x: Int, y: Int)
+    fun onScroll(x: Float, y: Float)
     {
         if (this is HorizontallyScrollable) xScroll = x
         if (this is VerticallyScrollable) yScroll = y
@@ -11,7 +11,7 @@ sealed interface Scrollable
 
 interface VerticallyScrollable : Scrollable
 {
-    var yScroll: Int
+    var yScroll: Float
     var verticalScrollbarVisibility: ScrollbarVisibility
     fun setVerticalScroll(fraction: Float)
     fun getVerticallyUsedSpaceFraction(): Float
@@ -19,7 +19,7 @@ interface VerticallyScrollable : Scrollable
 
 interface HorizontallyScrollable : Scrollable
 {
-    var xScroll: Int
+    var xScroll: Float
     var horizontalScrollbarVisibility: ScrollbarVisibility
     fun setHorizontalScroll(fraction: Float)
     fun getHorizontallyUsedSpaceFraction(): Float
